Siavash Malekpour Branch
Took a bit of a different approach than requested. The idea is that, the travelers can view their travel detail "after" they have traveled along with the link to other fellow travel members, destinations they've visited, and pictures and videos of their travel.

The UI is still yet to come